// src/types/apis/linkedin/TProfile.ts
export type GlobalProfile = {
  name: string;
  imageUrl: string;
  location: string;
  headline: string;
  totalcount: any;
  public_Identifier: string;
  entityUrn: string;
};

export type DropDownProfiles = {
  title: string;
  subtitle: string;
  image: string;
  entityUrn: string;
  publicIdentifier: string;
};